django/requirements.txt
Django
algoliasearch
djangorestframework
python-dotenv